Moved another non-test to test/unit/mock/files
