+++ test/unit/resources/aws_billing_report_test.rb
@@ -2,7 +2,7 @@ require "helper"
 require "inspec/resource"
 require "resources/aws/aws_billing_report"
 
-require_relative "aws_billing_backend" # TODO: move this to mocks
+require "unit/mock/files/aws_billing_backend"
 
 require "resource_support/aws"
 require "resources/aws/aws_billing_report"
