+++ test/unit/resources/aws_billing_reports_test.rb
@@ -2,7 +2,7 @@ require "helper"
 require "inspec/resource"
 require "resources/aws/aws_billing_reports"
 
-require_relative "aws_billing_backend"
+require "unit/mock/files/aws_billing_backend"
 
 require "resource_support/aws"
 require "resources/aws/aws_billing_reports"
